pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "speech-to-text-app"
version = "0.1.0"
description = "A speech-to-text converter using OpenAI's Whisper API"
readme = "README.md"
requires-python = ">=3.6"
authors = [
    {name = "Your Name"}
]
dependencies = [
    "openai>=1.3.0",
    "python-dotenv>=0.19.0",
    "keyboard>=0.13.5",
    "pystray>=0.19.4",
    "pillow>=9.0.0",
    "pyperclip>=1.8.2",
    "sounddevice>=0.4.5",
    "numpy>=1.22.0",
    "scipy>=1.8.0",
    "matplotlib>=3.5.0",
]

[project.optional-dependencies]
dev = [
    "black",
    "flake8",
    "pytest",
]

[tool.setuptools]
packages = ["speech_to_text"]
